requirements.txt
coverage
django-suit
South
django-userena
django-guardian
django-mailgun